+++ block.h
@@ -7,15 +7,15 @@ class Block : public QObject, public QGraphicsRectItem
 {
     Q_OBJECT
 public:
-    explicit Block(qreal x, qreal y, QColor &color, QGraphicsItem *parent = 0);
+    explicit Block(int x, int y, QColor &color, QGraphicsItem *parent = 0);
 
-    static const QPointF getPoint(const qreal &x, const qreal &y);
+    static const QPointF getPoint(const int &x, const int &y);
 
     static const int length = 10;
     static const int padding = 2;
 
 signals:
-    void hover(qreal x, qreal y);
+    void hover(int x, int y);
     void hoverEnd();
 
 private:
